+++ client/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Add filter by year on analysis table. [LANDGRIF-965](https://vizzuality.atlassian.net/browse/LANDGRIF-965)
 
 ### Changed
+- Signup removed from auth page.[LANDGRIF-1314](https://vizzuality.atlassian.net/browse/LANDGRIF-1314) 
 - Updated date and time format in the upload data page, also added the author [LANDGRIF-1297](https://vizzuality.atlassian.net/browse/LANDGRIF-1297)
 - Changed table style. [LANDGRIF-1299](https://vizzuality.atlassian.net/browse/LANDGRIF-1299)
 - Updated axios version.